Reformatted all methods to a new file Functions
